refactor: Refactor and package codes
