+++ src/project_2/channel_scenario.py
@@ -1,9 +1,9 @@
 import os
 
-from account import XrplAccount
-from transaction import XrplTransaction
-from request import XrplRequest
-from utils import Logger
+from src.xrpl_wrapper.account import XrplAccount
+from src.xrpl_wrapper.transaction import XrplTransaction
+from src.xrpl_wrapper.request import XrplRequest
+from src.xrpl_wrapper.utils import Logger
 
 DIR = os.path.dirname(__file__)
